cast to int

+++ scripts/experiments_settings.py
@@ -12,7 +12,7 @@ number_of_omp_threads = [1]
 number_of_training_samples = [10 + i * 20 for i in range(3)] 
 number_of_training_iterations = [1000 + i * 2000 for i in range(3)]
 
-number_of_initial_hidden_units = [q*(q-1)/2.0 for q in number_of_qubits]
+number_of_initial_hidden_units = [int(q*(q-1)/2.0) for q in number_of_qubits]
 number_of_sample_steps = [q if q%2 != 0 else q+1 for q in number_of_qubits]
 number_of_runs = 1 #number of runs for a specific circuit
